Clear out all CDAPI references
