+++ .env.test
@@ -5,9 +5,6 @@ COURT_DATA_ADAPTOR_API_TEST_MODE: true
 
 GA_TRACKING_ID: UA-XXXXXXXXX-XX
 
-# CDAPI
-COURT_DATA_API_URL: http://localhost:8000/v2
-
 # Enable by default the following features on test environment
 # once all are fixed we can remove them
