--- app/models/cd_api/cracked_ineffective_trial.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module CdApi
-  class CrackedIneffectiveTrial < BaseModel
-    belongs_to :hearing
-  end
-end
