--- app/models/cd_api/prosecution_case.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module CdApi
-  class ProsecutionCase < BaseModel
-    belongs_to :hearing, class_name: 'CdApi::Hearing'
-  end
-end
