--- app/views/prosecution_cases/_cracked_ineffective_trial.html.haml
@@ -1,14 +1,0 @@
-- if prosecution_case&.cracked?
-  %table.govuk-table
-    %caption.govuk-table__caption.govuk-visually-hidden
-      = t('prosecution_case.show.cracked_ineffective_trial.caption')
-    %tbody.govuk-table__body
-      %tr.govuk-table__row
-        %th.govuk-table__header{ scope: 'row' }
-          = t('prosecution_case.show.cracked_ineffective_trial.heading')
-        %td.govuk-table__cell
-          - prosecution_case.hearings.each do |hearing|
-            - next unless hearing.cracked_ineffective_trial&.cracked?
-            = hearing.cracked_ineffective_trial.cracked_on_sentence(hearing)
-            %br
-            = hearing.cracked_ineffective_trial.description
