--- app/views/prosecution_cases/_defendants.html.haml
@@ -1,19 +1,0 @@
-%h2.govuk-heading-l{ class: "govuk-!-margin-bottom-4" }
-  = t('search.result.defendant.heading')
-%table.govuk-table{ class: "govuk-!-margin-bottom-9" }
-  %caption.govuk-table__caption.govuk-visually-hidden
-    = t('search.result.defendant.caption')
-  %thead.govuk-table__head
-    %tr.govuk-table__row
-      %th.govuk-table__header{ scope: 'col' }= t('search.result.defendant.name')
-      %th.govuk-table__header{ scope: 'col' }= t('search.result.defendant.date_of_birth')
-      %th.govuk-table__header{ scope: 'col' }= t('search.result.defendant.maat_number')
-  %tbody.govuk-table__body
-    - prosecution_case.defendants.each do |defendant|
-      %tr.govuk-table__row
-        %td.govuk-table__cell
-          = link_to defendant.name, defendant_link_path(defendant, prosecution_case.prosecution_case_reference), class: 'govuk-link'
-        %td.govuk-table__cell
-          = l(defendant.date_of_birth&.to_date)
-        %td.govuk-table__cell
-          = defendant.maat_reference || t('search.result.defendant.unlinked')
