+++ app/views/prosecution_cases/_hearing_summaries.html.haml
