+++ spec/features/hearings_spec.rb
@@ -3,11 +3,11 @@
 RSpec.feature 'Viewing the hearings page', :stub_case_search, :stub_v2_hearing_summary,
               :stub_internal_v2_hearing_results, type: :feature do
   let(:user) { create(:user) }
-  let(:api_url_v2) { CdApi::BaseModel.site }
-  let(:api_events_path) { "#{Cda::BaseModel.site}hearings/#{hearing_id}/event_log/2019-10-23" }
-  let(:api_data_path) { "#{Cda::BaseModel.site}hearing_results/#{hearing_id}" }
+  let(:api_url_v2) { Cda::BaseModel.site }
+  let(:api_events_path) { "#{api_url_v2}hearings/#{hearing_id}/event_log/2019-10-23" }
+  let(:api_data_path) { "#{api_url_v2}hearing_results/#{hearing_id}" }
   let(:case_reference) { 'TEST12345' }
-  let(:api_summary_path) { "#{Cda::BaseModel.site}prosecution_cases" }
+  let(:api_summary_path) { "#{api_url_v2}prosecution_cases" }
   let(:hearing_id) { '345be88a-31cf-4a30-9de3-da98e973367e' }
 
   before do
