--- spec/fixtures/stubs/cd_api/hearing_summary_response.json
@@ -1,191 +1,0 @@
-{
-  "prosecution_case_reference": "TEST12345",
-  "hearing_summaries": [
-    {
-      "id": "ffebc761-1b61-4ba5-b197-c7326d002d92",
-      "hearing_type": "Trial (TRL)",
-      "estimated_duration": null,
-      "defendants": [],
-      "jurisdiction_type": "MAGISTRATES",
-      "court_centre": {
-        "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-        "name": "Derby Crown Court",
-        "welsh_name": "Llys Y Goron Derby",
-        "room_id": null,
-        "room_name": null,
-        "welsh_room_name": null,
-        "welsh_court_name": null,
-        "short_oucode": "C30DE",
-        "oucode_l2_code": "30",
-        "code": "C30DE00",
-        "address": {
-          "address_1": "Derby Combined Court Centre",
-          "address_2": "Morledge",
-          "address_3": "Derby",
-          "address_4": "",
-          "address_5": "",
-          "postcode": "DE1 2XE"
-        }
-      },
-      "hearing_days": [
-        {
-          "sitting_day": "2019-10-23T08:30:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-24T08:45:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-25T09:00:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        }
-      ],
-      "defence_counsels": []
-    },
-    {
-      "id": "c2002ca4-fa81-4639-9b7c-58940bfbe6f5",
-      "hearing_type": "Mention - Defendant to Attend (MDA)",
-      "estimated_duration": null,
-      "defendants": [],
-      "jurisdiction_type": "CROWN",
-      "court_centre": {
-        "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-        "name": "Derby Crown Court",
-        "welsh_name": "Llys Y Goron Derby",
-        "room_id": null,
-        "room_name": null,
-        "welsh_room_name": null,
-        "welsh_court_name": null,
-        "short_oucode": "C30DE",
-        "oucode_l2_code": "30",
-        "code": "C30DE00",
-        "address": {
-          "address_1": "Derby Combined Court Centre",
-          "address_2": "Morledge",
-          "address_3": "Derby",
-          "address_4": "",
-          "address_5": "",
-          "postcode": "DE1 2XE"
-        }
-      },
-      "hearing_days": [
-        {
-          "sitting_day": "2019-10-26T08:30:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-27T08:45:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-28T09:00:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        }
-      ],
-      "defence_counsels": []
-    },
-    {
-      "id": "345be88a-31cf-4a30-9de3-da98e973367e",
-      "hearing_type": "Pre-Trial Review (PTR)",
-      "estimated_duration": null,
-      "defendants": [],
-      "jurisdiction_type": "CROWN",
-      "court_centre": {
-        "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-        "name": "Derby Crown Court",
-        "welsh_name": "Llys Y Goron Derby",
-        "room_id": null,
-        "room_name": null,
-        "welsh_room_name": null,
-        "welsh_court_name": null,
-        "short_oucode": "C30DE",
-        "oucode_l2_code": "30",
-        "code": "C30DE00",
-        "address": {
-          "address_1": "Derby Combined Court Centre",
-          "address_2": "Morledge",
-          "address_3": "Derby",
-          "address_4": "",
-          "address_5": "",
-          "postcode": "DE1 2XE"
-        }
-      },
-      "hearing_days": [
-        {
-          "sitting_day": "2019-10-29T08:30:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-30T08:45:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        },
-        {
-          "sitting_day": "2019-10-31T09:00:00+00:00",
-          "listing_sequence": 1,
-          "listed_duration_minutes": 1,
-          "has_shared_results": null
-        }
-      ],
-      "defence_counsels": []
-    }
-  ],
-  "overall_defendants": [
-    {
-      "id": "a23d0911-4806-42eb-b910-e2bdcf3018d9",
-      "first_name": "Wendie",
-      "middle_name": "Bogisiche",
-      "last_name": "Lowe",
-      "arrest_summons_number": "FPV3XYUJNPIB",
-      "date_of_birth": "2000-12-27",
-      "national_insurance_number": "HX685369B",
-      "maat_reference": "Z10000233"
-    },
-    {
-      "id": "018c6466-ffc3-4020-8b33-1d6702b4d189",
-      "first_name": "Marlin",
-      "middle_name": "Schaefer",
-      "last_name": "Leuschke",
-      "arrest_summons_number": "OYOYFT1DFB4S",
-      "date_of_birth": "1958-05-08",
-      "national_insurance_number": "YZ235772C",
-      "maat_reference": "Z10000204"
-    },
-    {
-      "id": "66ef1647-c5e3-4b10-a484-0227afb116fb",
-      "first_name": "Randal",
-      "middle_name": "Huels",
-      "last_name": "Mitchell",
-      "arrest_summons_number": "5SLSOIL9Q2ZQ",
-      "date_of_birth": "1967-09-24",
-      "national_insurance_number": "AH506781B",
-      "maat_reference": "Z10000178"
-    },
-    {
-      "id": "8b7a9989-84ba-493d-8780-a97465872c53",
-      "first_name": "Jammy",
-      "middle_name": "",
-      "last_name": "Dodger",
-      "arrest_summons_number": "1PBAP0Q18V3F",
-      "date_of_birth": "1987-05-21",
-      "national_insurance_number": "JC123456A",
-      "maat_reference": "Z10000222"
-    }
-  ]
-}
