--- spec/fixtures/stubs/cd_api/hearingsummaries_body.json
@@ -1,191 +1,0 @@
-{
-  "prosecution_case_reference": "TEST12345",
-  "hearing_summaries": [
-      {
-          "id": "8e267333-1e15-438f-ad31-1c8b89f1a919",
-          "hearing_type": "Trial (TRL) TEST",
-          "estimated_duration": null,
-          "defendants": [],
-          "jurisdiction_type": null,
-          "court_centre": {
-              "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "name": "Derby Crown Court",
-              "welsh_name": "Llys Y Goron Derby",
-              "room_id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "room_name": "Derby Crown Court",
-              "welsh_room_name": "Llys Y Goron Derby",
-              "welsh_court_name": null,
-              "short_oucode": "C30DE",
-              "oucode_l2_code": "30",
-              "code": "C30DE00",
-              "address": {
-                  "address_1": "Derby Combined Court Centre",
-                  "address_2": "Morledge",
-                  "address_3": "Derby",
-                  "address_4": null,
-                  "address_5": null,
-                  "postcode": "DE1 2XE"
-              }
-          },
-          "hearing_days": [
-              {
-                  "sitting_day": "2019-10-23T08:30:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-24T08:45:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-25T09:00:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              }
-          ],
-          "defence_counsels": []
-      },
-      {
-          "id": "95c3c65f-a58e-44c0-8de6-a9a9949f93fb",
-          "hearing_type": "Pre-Trial Review (PTR) TEST",
-          "estimated_duration": null,
-          "defendants": [],
-          "jurisdiction_type": null,
-          "court_centre": {
-              "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "name": "Derby Crown Court",
-              "welsh_name": "Llys Y Goron Derby",
-              "room_id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "room_name": "Derby Crown Court",
-              "welsh_room_name": "Llys Y Goron Derby",
-              "welsh_court_name": null,
-              "short_oucode": "C30DE",
-              "oucode_l2_code": "30",
-              "code": "C30DE00",
-              "address": {
-                  "address_1": "Derby Combined Court Centre",
-                  "address_2": "Morledge",
-                  "address_3": "Derby",
-                  "address_4": null,
-                  "address_5": null,
-                  "postcode": "DE1 2XE"
-              }
-          },
-          "hearing_days": [
-              {
-                  "sitting_day": "2019-10-26T08:30:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-27T08:45:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-28T09:00:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              }
-          ],
-          "defence_counsels": []
-      },
-      {
-          "id": "b5cdd583-ec45-4599-bebc-586684c7b998",
-          "hearing_type": "Mention - Defendant to Attend (MDA) TEST",
-          "estimated_duration": null,
-          "defendants": [],
-          "jurisdiction_type": null,
-          "court_centre": {
-              "id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "name": "Derby Crown Court",
-              "welsh_name": "Llys Y Goron Derby",
-              "room_id": "6131bd34-33d9-3d1e-8152-8b5a2084f1bd",
-              "room_name": "Derby Crown Court",
-              "welsh_room_name": "Llys Y Goron Derby",
-              "welsh_court_name": null,
-              "short_oucode": "C30DE",
-              "oucode_l2_code": "30",
-              "code": "C30DE00",
-              "address": {
-                  "address_1": "Derby Combined Court Centre",
-                  "address_2": "Morledge",
-                  "address_3": "Derby",
-                  "address_4": null,
-                  "address_5": null,
-                  "postcode": "DE1 2XE"
-              }
-          },
-          "hearing_days": [
-              {
-                  "sitting_day": "2019-10-29T08:30:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-30T08:45:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              },
-              {
-                  "sitting_day": "2019-10-31T09:00:00+00:00",
-                  "listing_sequence": 1,
-                  "listed_duration_minutes": 1,
-                  "has_shared_results": null
-              }
-          ],
-          "defence_counsels": []
-      }
-  ],
-  "overall_defendants": [
-      {
-          "id": "29d1db1c-1467-4304-aa5b-904044a1afe3",
-          "first_name": "Lakesha",
-          "middle_name": "Hirthe",
-          "last_name": "Metz",
-          "arrest_summons_number": "R74BJ3VVK7BE",
-          "date_of_birth": "1978-12-10",
-          "national_insurance_number": "XE123410C",
-          "maat_reference": null
-      },
-      {
-          "id": "28d6f55e-2c74-4e61-a634-666718561cbf",
-          "first_name": "Jordan",
-          "middle_name": "O'Conner",
-          "last_name": "Rosenbaum",
-          "arrest_summons_number": "3A21EIGYRZW3",
-          "date_of_birth": "1981-03-17",
-          "national_insurance_number": "GG325302A",
-          "maat_reference": null
-      },
-      {
-          "id": "dec4649c-744a-420b-9647-bd7f5ca34ae9",
-          "first_name": "Meagan",
-          "middle_name": "Turcotte",
-          "last_name": "Johns",
-          "arrest_summons_number": "SMP0ZLSQJ9P3",
-          "date_of_birth": "1991-10-18",
-          "national_insurance_number": "RR856775C",
-          "maat_reference": null
-      },
-      {
-          "id": "b132aba1-5cbd-47ff-88da-639a046ddcb2",
-          "first_name": "Jammy",
-          "middle_name": "",
-          "last_name": "Dodger",
-          "arrest_summons_number": "ZNLLJPANVQBS",
-          "date_of_birth": "1987-05-21",
-          "national_insurance_number": "JC123456A",
-          "maat_reference": null
-      }
-  ]
-}
