+++ spec/fixtures/stubs/cda/hearing_data_empty_response.json
