+++ spec/fixtures/stubs/cda/hearing_data_response.json
