+++ spec/fixtures/stubs/cda/hearing_events_empty_response.json
