+++ spec/fixtures/stubs/cda/hearing_events_response.json
